==> 07-manual_resell.ipynb <==
import glob
import os.path

from include.global_functions import init_driver
from include.cbfunctions import *

folder_path = 'db\out'
file_type = '/*.csv'
files = glob.glob(folder_path + file_type)
max_file = max(files, key=os.path.getctime)

max_file

top_offer_df = pd.read_csv(max_file)
top_offer_df = top_offer_df.head(1)
top_offer_df

driver = init_driver()

#buy_boat(driver, top_offer_df, 2)

#sell_boat(driver, acc_id=2, resale_discount=99, price=0.04)
#sell_boat(driver, acc_id=1, resale_discount=5)

resell_auction(driver=driver, acc_id=2, resale_discount=99, price=0.04)

#resell_auction_pct(driver, acc_id=1, pct_resale=0)

driver.quit()
