==> 20-extract_misc.ipynb <==
#USED FOR AIRFLOW
import os
try:
    relative_location
except NameError:
    relative_location = ""

if relative_location != "":
    os.chdir(relative_location)

import pandas as pd
import csv
import time
import requests
from os.path import exists
from datetime import date, datetime, timezone

sold_stats_url = f'https://api.cryptobay.io/bay/cryptobaygetauctionsummary?data=%7B%22timestamp%22%3A{int(time.time())}%7D'
#sold_stats_url = f'https://api.cryptobay.top/bay/cryptobaygetauctionsummary?data=%7B%22timestamp%22%3A{int(time.time())}%7D'
raw_json = requests.get(sold_stats_url).json()
raw_json
columns = ['date', 'time', 'total_sale_24h', 'total_sold_24h',\
    'total_sale_7d', 'total_sold_7d', 'total_sale_31d', 'total_sold_31d']

day = datetime.today().strftime('%Y-%m-%d')
time = datetime.now(timezone.utc).strftime('%H:%M:%S')
total_sale_24h = raw_json['data']['24h']['total_sale']
total_sold_24h = raw_json['data']['24h']['total_sold']

total_sale_7d = raw_json['data']['7days']['total_sale']
total_sold_7d = raw_json['data']['7days']['total_sold']

total_sale_31d = raw_json['data']['30days']['total_sale']
total_sold_31d = raw_json['data']['30days']['total_sold']

# Create file if it doesn't exist
sold_stats_file = "db/misc/total_sold_stats.csv"
sold_stats_df = pd.DataFrame(columns=columns)

# If file exists, then append extracted data
_series = pd.Series([day, time, total_sale_24h, total_sold_24h,\
        total_sale_7d, total_sold_7d, total_sale_31d, total_sold_31d], index=columns)
sold_stats_df = sold_stats_df.append(_series, ignore_index=True)
sold_stats_df.to_csv(sold_stats_file, index=False, mode='a', header=False)
