==> 70-EDA.ipynb <==
# TO DO  GENERATOR
import random
random.seed(1)

def r_seq(n):
    x = [i for i in random.random()]
    return x 

r_seq(20)

# TO DO 
[ i for i in range(5) if i==3 or i==4]

# TO DO 
x = (1, 2, 3)
s = set(x)
s
# {1, 2, 3}

# TO DO
x = lambda : "I know how to call this function."
x()

# TO DO
#The count_words() function takes a text file as input and prints the number of words contained in the text. The function should print an error message when the file cannot be found. The number of words in the file should be printed if the file can be found.
def count_words(filename):
  """Count the number of words in a file"""
  
___ :
    with open(filename, encoding='utf-8') as f:
      contents = f.read()
  
___ FileNotFoundError:
    print("The file {} cannot be found".format(filename))
  else:
    words = contents.split()
    num_words = len(words)
    print("The file {} has about {} words.".format(filename, num_words))
    
filename = "peter_rabbit.txt"
count_words(filename)
